get user posts in tab 2
